requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
apscheduler==3.10.4
pydantic==2.9.2
python-dotenv==1.0.1
